fix: split logic in different files
